rename: change a file name notation - from: Pascal Case - to: Kebab Case
