+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Route, Routes } from "react-router";
 import { BrowserRouter } from "react-router-dom";
-import Detail from "./routers/Detail";
-import Home from "./routers/Home";
-import Volunteer from "./routers/Volunteer";
+import Detail from "./routers/detail";
+import Home from "./routers/home";
+import Volunteer from "./routers/volunteer";
 import "./App.css";
 
 const App = () => {
