+++ src/components/logo-slide.tsx
