+++ src/routers/Volunteer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Footer from "../components/common/Footer";
-import Header from "../components/common/Header";
-import VolunteerList from "../components/VolunteerList";
+import Footer from "../components/common/footer";
+import Header from "../components/common/header";
+import VolunteerList from "../components/volunteer-list";
 
 const Volunteer = () => {
   return (
